+++ pages/api/posts.js
@@ -76,14 +76,7 @@ async function deletePost(req, res) {
         _id: {
           $in: idList
         }
-      },
-        function (err, result) {
-          if (err) {
-            res.send(err);
-          } else {
-            res.send(result);
-          }
-        });
+      });
     } else {
       await db.collection(table).deleteOne({
         _id: ObjectId(reqBody.id),